Move Store to own file.
